+++ tests/narwhals_ext/test_narwhals_dataframe.py
@@ -4,6 +4,7 @@ import pytest
 
 import laktory as lk  # noqa: F401
 from laktory._testing import StreamingSource
+from laktory._testing import get_df0
 
 
 @pytest.fixture()
@@ -38,9 +39,33 @@ def df0():
     )
 
 
+@pytest.mark.parametrize(
+    ("backend", "is_streaming"),
+    [
+        ("POLARS", False),
+        ("PYSPARK", False),
+        ("PYSPARK", True),
+    ],
+)
+def test_dispaly(backend, is_streaming, tmp_path):
+    if not is_streaming:
+        df = get_df0(backend=backend)
+    else:
+        spark = lk.get_spark_session()
+
+        ss = StreamingSource(backend)
+        ss.write_to_delta(str(tmp_path))
+        df = spark.readStream.format("DELTA").load(str(tmp_path))
+        print(df.isStreaming)
+        df = nw.from_native(df)
+
+    # TODO: capture stdout to test output
+    df.laktory.display(n=2, timeout=1)
+
+
 @pytest.mark.parametrize("backend", ["POLARS", "PYSPARK"])
 def test_groupby_and_agg(backend):
-    ss = StreamingSource()
+    ss = StreamingSource(backend=backend)
     df0 = nw.concat(ss.get_dfs(3)).collect()
 
     df = df0.laktory.groupby_and_agg(
@@ -63,37 +88,6 @@ def test_has_column(df0):
     assert df0.laktory.has_column("x@x")
     assert df0.laktory.has_column("`x@x`")
     assert df0.laktory.has_column("u[0].a")
-
-
-def test_schema_flat(df0):
-    schema = df0.laktory.schema_flat()
-    assert schema == [
-        "x@x",
-        "y",
-        "z",
-        "z.id",
-        "z.email",
-        "u",
-        "u[*].a",
-        "u[*].b",
-        "u2",
-        "u2[*].a",
-        "u2[*].b",
-    ]
-
-
-def test_union(df0):
-    df = df0.laktory.union(df0)
-    assert df.shape[0] == df0.shape[0] * 2
-    assert df.schema == df0.schema
-
-
-def test_signature(df0):
-    sig = df0.select("x@x", "y", "z").laktory.signature()
-    assert (
-        sig
-        == "DataFrame[x@x: Int64, y: List(String), z: Struct({'id': Int64, 'email': String})]"
-    )
 
 
 @pytest.mark.parametrize("is_lazy", [True, False])
@@ -125,6 +119,37 @@ def test_row_index(is_lazy):
         assert df.sort("i0")["i0"].to_list() == list(range(9))
     assert df.sort("y")["i1"].to_list() == list(range(9))
     assert df.sort("x", "y")["i2"].to_list() == [0, 1, 2] * 3
+
+
+def test_schema_flat(df0):
+    schema = df0.laktory.schema_flat()
+    assert schema == [
+        "x@x",
+        "y",
+        "z",
+        "z.id",
+        "z.email",
+        "u",
+        "u[*].a",
+        "u[*].b",
+        "u2",
+        "u2[*].a",
+        "u2[*].b",
+    ]
+
+
+def test_signature(df0):
+    sig = df0.select("x@x", "y", "z").laktory.signature()
+    assert (
+        sig
+        == "DataFrame[x@x: Int64, y: List(String), z: Struct({'id': Int64, 'email': String})]"
+    )
+
+
+def test_union(df0):
+    df = df0.laktory.union(df0)
+    assert df.shape[0] == df0.shape[0] * 2
+    assert df.schema == df0.schema
 
 
 def test_window_filter():